+++ src/core_modules/capture-core/components/WidgetEventEdit/EditEventDataEntry/EditEventDataEntry.component.js
@@ -1,6 +1,7 @@
 // @flow
 import React, { Component } from 'react';
 import { withStyles } from '@material-ui/core/styles';
+import { dataEntryIds } from 'capture-core/constants';
 import { TabBar, Tab } from '@dhis2/ui';
 import i18n from '@dhis2/d2-i18n';
 import type { OrgUnit } from 'capture-core-utils/rulesEngine';
@@ -32,10 +33,11 @@ import {
     withDefaultFieldContainer,
     withDefaultShouldUpdateInterface,
 } from '../../FormFields/New';
-import { statusTypes } from '../../../events/statusTypes';
+import { statusTypes, translatedStatusTypes } from '../../../events/statusTypes';
 import { inMemoryFileStore } from '../../DataEntry/file/inMemoryFileStore';
 import labelTypeClasses from '../DataEntry/dataEntryFieldLabels.module.css';
 import { withDeleteButton } from '../DataEntry/withDeleteButton';
+import { withAskToCreateNew } from '../../DataEntry/withAskToCreateNew';
 import { actionTypes } from './editEventDataEntry.actions';
 
 const tabMode = Object.freeze({
@@ -140,7 +142,7 @@ const buildReportDateSettingsFn = () => {
 };
 
 const buildScheduleDateSettingsFn = () => {
-    const scheduleDateComponent =
+    const scheduleDateComponent = innerProps =>
         withCalculateMessages(overrideMessagePropNames)(
             withFocusSaver()(
                 withDefaultFieldContainer()(
@@ -149,7 +151,15 @@ const buildScheduleDateSettingsFn = () => {
                             onGetUseVerticalOrientation: (props: Object) => props.formHorizontal,
                             onGetCustomFieldLabeClass: (props: Object) =>
                                 `${props.fieldOptions.fieldLabelMediaBasedClass} ${labelTypeClasses.dateLabel}`,
-                            customTooltip: i18n.t('Go to “Schedule” tab to reschedule this event'),
+                            customTooltip: () => {
+                                const isScheduleableStatus =
+                                [statusTypes.SCHEDULE, statusTypes.OVERDUE].includes(innerProps.eventStatus);
+
+                                return isScheduleableStatus ?
+                                    i18n.t('Go to “Schedule” tab to reschedule this event') :
+                                    i18n.t('Scheduled date cannot be changed for {{ eventStatus }} events',
+                                        { eventStatus: translatedStatusTypes()[innerProps.eventStatus] });
+                            },
                         })(
                             withDisplayMessages()(
                                 withInternalChangeHandler()(withFilterProps(defaultFilterProps)(DateField)),
@@ -160,14 +170,14 @@ const buildScheduleDateSettingsFn = () => {
             ),
         );
     const scheduleDateSettings = {
-        getComponent: () => scheduleDateComponent,
+        getComponent: (props: Object) => scheduleDateComponent(props),
         getComponentProps: (props: Object) => createComponentProps(props, {
             width: '100%',
             calendarWidth: 350,
             label: props.formFoundation.getLabel('scheduledAt'),
             disabled: true,
         }),
-        getIsHidden: (props: Object) => ![statusTypes.SCHEDULE, statusTypes.OVERDUE].includes(props.eventStatus),
+        getIsHidden: (props: Object) => props.id !== dataEntryIds.ENROLLMENT_EVENT || props.hideDueDate,
         getPropName: () => 'scheduledAt',
         getValidatorContainers: () => getEventDateValidatorContainers(),
         getMeta: () => ({
@@ -304,12 +314,14 @@ const SaveableDataEntry = withSaveHandler(saveHandlerConfig)(withMainButton()(Re
 const CancelableDataEntry = withCancelButton(getCancelOptions)(SaveableDataEntry);
 const CompletableDataEntry = withDataEntryField(buildCompleteFieldSettingsFn())(CancelableDataEntry);
 const DeletableDataEntry = withDeleteButton()(CompletableDataEntry);
-const DataEntryWrapper = withBrowserBackWarning()(DeletableDataEntry);
+const AskToCreateNewDataEntry = withAskToCreateNew()(DeletableDataEntry);
+const DataEntryWrapper = withBrowserBackWarning()(AskToCreateNewDataEntry);
 
 type Props = {
     formFoundation: ?RenderFoundation,
     orgUnit: OrgUnit,
     programId: string,
+    itemId: string,
     initialScheduleDate?: string,
     onUpdateDataEntryField: (orgUnit: OrgUnit, programId: string) => (innerAction: ReduxAction<any, any>) => void,
     onUpdateField: (orgUnit: OrgUnit, programId: string) => (innerAction: ReduxAction<any, any>) => void,
@@ -318,6 +330,8 @@ type Props = {
     onHandleScheduleSave: (eventData: Object) => void,
     onDelete: () => void,
     onCancel: () => void,
+    onConfirmCreateNew: (itemId: string) => void,
+    onCancelCreateNew: (itemId: string) => void,
     classes: {
         dataEntryContainer: string,
         fieldLabelMediaBased?: ?string,
@@ -327,6 +341,7 @@ type Props = {
     onCancelEditEvent?: () => void,
     eventStatus?: string,
     enrollmentId?: string,
+    isCompleted?: boolean,
 };
 
 type State = {
@@ -427,6 +442,7 @@ class EditEventDataEntryPlain extends Component<Props, State> {
             classes,
             ...passOnProps
         } = this.props;
+
         return ( // $FlowFixMe[cannot-spread-inexact] automated comment
             <DataEntryWrapper
                 id={dataEntryId}