+++ src/core_modules/capture-core/components/WidgetFeedback/WidgetFeedbackContent/WidgetFeedbackContent.js
@@ -1,7 +1,7 @@
 // @flow
 
 import React, { type ComponentType } from 'react';
-import { spacers, spacersNum, colors } from '@dhis2/ui';
+import { spacersNum, colors } from '@dhis2/ui';
 import { withStyles } from '@material-ui/core';
 import type { FilteredKeyValue, FilteredText, ContentType, WidgetData } from '../WidgetFeedback.types';
 
@@ -11,22 +11,25 @@ const styles = {
     },
     unorderedList: {
         paddingLeft: spacersNum.dp16,
-        marginTop: '0px',
-        lineHeight: '1.375',
-        fontSize: spacers.dp16,
+        marginTop: 0,
+        fontSize: '14px',
+        lineHeight: '19px',
         color: colors.grey900,
     },
     noFeedbackText: {
         color: colors.grey600,
-        paddingLeft: spacersNum.dp16,
         fontWeight: 400,
-        fontSize: spacers.dp16,
-        marginTop: '0px',
+        fontSize: '14px',
+        lineHeight: '19px',
+        marginTop: 0,
     },
     listItem: {
-        marginBottom: spacersNum.dp4,
+        marginBottom: spacersNum.dp8,
         '&::marker': {
             color: colors.grey500,
+        },
+        '& + &': {
+            marginBottom: 0,
         },
     },
 };
@@ -34,7 +37,14 @@ const styles = {
 
 const WidgetFeedbackContentComponent = ({ widgetData, emptyText, classes }: ContentType) => {
     if (!widgetData?.length) {
-        return <p data-test="widget-content" className={classes.noFeedbackText}>{emptyText}</p>;
+        return (
+            <div
+                data-test="widget-content"
+                className={classes.container}
+            >
+                <p className={classes.noFeedbackText}>{emptyText}</p>
+            </div>
+        );
     }
 
     const renderTextObject = (item: FilteredText) => (